+++ vite-project/src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import ProjectDetail from "./pages/ProjectDetail";
+import ProjectDashboard from "./pages/ProjectDashboard";
 import "./App.css";
 import Home from "./pages/Home";
 
@@ -9,7 +9,7 @@ const App = () => {
     <Router>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/project/:address" element={<ProjectDetail />} />
+        <Route path="/project/:address" element={<ProjectDashboard />} />
       </Routes>
     </Router>
   );
