+++ vite-project/src/components/Loader.jsx
@@ -1,25 +1,3 @@
-// import { Box } from "@mui/material";
-// import Backdrop from "@mui/material/Backdrop";
-// import CircularProgress from "@mui/material/CircularProgress";
-
-// export default function Loader() {
-//   return (
-//     <Box>
-//       <Backdrop
-//         sx={{
-//           color: "#fff",
-//           zIndex: (theme) => theme.zIndex.drawer + 1,
-//           opacity: 0.5,
-//         }}
-//         open={true}
-//       >
-//         <CircularProgress color="inherit" />
-//       </Backdrop>
-//     </Box>
-//   );
-// }
-
-// Loader.jsx
 import { Typography } from "@mui/material";
 import Box from "@mui/material/Box";
 import CircularProgress from "@mui/material/CircularProgress";
